Add new autolinker
The new autolinker works on AST. Actually, two ASTs!

1. it takes HTML AST and returns updated AST. This way, we're only
   trying to autolink `{:code, _, _}` nodes (but not e.g. `{:pre, _,
   [{:code, _, _}]}` etc.)

2. the Elixir AST. Turns out `Foo.bar/1` is actually valid Elixir code
   so we can use `Code.string_to_quoted/1` to parse that. This way we
   reliably recognize local & remote functions, special forms etc.

   This means we could trivially autolink _expressions_ e.g.:

       `String.upcase("foo")`

   but support for this is not on the roadmap at the moment.

We're also changing how we manage cross-references data. The new
(private) `ExDoc.Refs` module maintains a cache of references
and we can ask it questions like:

    ExDoc.Refs.public?({:module, String}) #=> true
    ExDoc.Refs.public?({:module, String.Unicode}) #=> false
    ExDoc.Refs.public?({:function, String, :upcase, 2}) #=> true
    # etc

If it hasn't yet seen a given module it tries to fetch it's docs chunk
and put it into the cache.

This means we have more precise references data. This is especially
useful for dependencies, previously we didn't check if remote calls
exist, e.g. referencing `String.upcase/9` succeeded, now it does not and
it warns.

Since refs work on chunks and OTP as of 23.0-rc1 can generate them, it
means we have better support for Erlang and other BEAM languages. This
is now autolinking:

    `c::gen_server.handle_call/3`
    `t::array.array/0`

We also extend custom links to support "erlang" modules:

    [See `:lists` module](`:lists`)